Adding runtime.txt, to bump the default python version
The default of 2.7.6 works, but triggers SSL warnings when some of the
newer versions of url fetching libraries are used.

In this case, this scraper is using `requests`, which in turn pulls in
the `urllib3` library. When using any python version below 2.7.9, `urllib3`
spits out this warning:

```
InsecurePlatformWarning: A true SSLContext object is not available. This
prevents urllib3 from configuring SSL appropriately and may cause
certain SSL connections to fail. For more information, see
https://urllib3.readthedocs.org/en/latest/security.html#insecureplatformwarning.
```